\name{MinLinLogistic-package}
\alias{MinLinLogistic-package}
\alias{MinLinLogistic}
\docType{package}
\title{
\packageTitle{MinLinLogistic}
}
\description{
\packageDescription{MinLinLogistic}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MinLinLogistic}
\packageIndices{MinLinLogistic}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MinLinLogistic}

Maintainer: \packageMaintainer{MinLinLogistic}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
